objects now use palette and bg priority flags

+++ fifo.go
@@ -3,7 +3,7 @@ package main
 type Pixel struct {
 	c          byte // 0-3
 	pal        byte // bit 4 OAM byte 3, 0=OBP0, 1=OBP1
-	bgPriority int  // bit 7 OAM byte 3, 0=obj above bg, 1=bg above obj
+	bgPriority byte  // bit 7 OAM byte 3, 0=obj above bg, 1=bg above obj
 }
 
 type FIFO []Pixel
